credit-postgresql/app/requirements.txt
psycopg2-binary==2.9.9
ddtrace>=2.8.0
